src/test/java/org/example/string/LongestSubstringWithoutRepeatingCharactersImplTest.java
package org.example.string;

import java.util.stream.Stream;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import static org.junit.jupiter.api.Assertions.*;

class LongestSubstringWithoutRepeatingCharactersImplTest
{
    @ParameterizedTest
    @MethodSource("provideTestData")
    void test(String str, int expectedResult)
    {
        int actualResult = LongestSubstringWithoutRepeatingCharactersImpl.lengthOfLongestSubstring(str);
        assertEquals(expectedResult, actualResult);
    }

    private static Stream<Arguments> provideTestData()
    {
        return Stream.of(
                Arguments.of(" ", 1),
                Arguments.of("abcabcbb", 3),
                Arguments.of("bbbbb", 1),
                Arguments.of("pwwkew", 3)
        );
    }

}